+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/ViewPagerActivity.kt
@@ -80,6 +80,7 @@ class ViewPagerActivity : SimpleActivity(), ViewPager.OnPageChangeListener, View
         var screenHeight = 0
     }
 
+    @TargetApi(Build.VERSION_CODES.P)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_medium)
@@ -336,6 +337,24 @@ class ViewPagerActivity : SimpleActivity(), ViewPager.OnPageChangeListener, View
                     }
                 }
             }
+        }
+
+        if (intent.action == "com.android.camera.action.REVIEW") {
+            Thread {
+                if (galleryDB.MediumDao().getMediaFromPath(mPath).isEmpty()) {
+                    val type = when {
+                        mPath.isVideoFast() -> TYPE_VIDEOS
+                        mPath.isGif() -> TYPE_GIFS
+                        mPath.isSvg() -> TYPE_SVGS
+                        mPath.isRawFast() -> TYPE_RAWS
+                        else -> TYPE_IMAGES
+                    }
+
+                    val duration = if (type == TYPE_VIDEOS) mPath.getVideoDuration() else 0
+                    val medium = Medium(null, mPath.getFilenameFromPath(), mPath, mPath.getParentPath(), System.currentTimeMillis(), System.currentTimeMillis(), File(mPath).length(), type, duration, false, 0)
+                    galleryDB.MediumDao().insert(medium)
+                }
+            }.start()
         }
     }
 
@@ -936,8 +955,16 @@ class ViewPagerActivity : SimpleActivity(), ViewPager.OnPageChangeListener, View
     }
 
     private fun askConfirmDelete() {
-        val message = if (config.useRecycleBin && !getCurrentMedium()!!.getIsInRecycleBin()) R.string.are_you_sure_recycle_bin else R.string.are_you_sure_delete
-        DeleteWithRememberDialog(this, getString(message)) {
+        val filename = "\"${getCurrentPath().getFilenameFromPath()}\""
+
+        val baseString = if (config.useRecycleBin && !getCurrentMedium()!!.getIsInRecycleBin()) {
+            R.string.move_to_recycle_bin_confirmation
+        } else {
+            R.string.deletion_confirmation
+        }
+
+        val message = String.format(resources.getString(baseString), filename)
+        DeleteWithRememberDialog(this, message) {
             config.tempSkipDeleteConfirmation = it
             deleteConfirmed()
         }