+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/helpers/Config.kt
@@ -423,4 +423,12 @@ class Config(context: Context) : BaseConfig(context) {
     var lastEditorCropOtherAspectRatioY: Int
         get() = prefs.getInt(LAST_EDITOR_CROP_OTHER_ASPECT_RATIO_Y, 1)
         set(lastEditorCropOtherAspectRatioY) = prefs.edit().putInt(LAST_EDITOR_CROP_OTHER_ASPECT_RATIO_Y, lastEditorCropOtherAspectRatioY).apply()
+
+    var groupDirectSubfolders: Boolean
+        get() = prefs.getBoolean(GROUP_DIRECT_SUBFOLDERS, false)
+        set(groupDirectSubfolders) = prefs.edit().putBoolean(GROUP_DIRECT_SUBFOLDERS, groupDirectSubfolders).apply()
+
+    var showWidgetFolderName: Boolean
+        get() = prefs.getBoolean(SHOW_WIDGET_FOLDER_NAME, true)
+        set(showWidgetFolderName) = prefs.edit().putBoolean(SHOW_WIDGET_FOLDER_NAME, showWidgetFolderName).apply()
 }