+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/helpers/MediaFetcher.kt
@@ -364,6 +364,10 @@ class MediaFetcher(val context: Context) {
                 else -> o1.taken.compareTo(o2.taken)
             }
 
+            if (result == 0) {
+                result = AlphanumericComparator().compare(o1.path.toLowerCase(), o2.path.toLowerCase())
+            }
+
             if (sorting and SORT_DESCENDING != 0) {
                 result *= -1
             }